Fixed UI locking when retrieving data from database.

+++ AuctionData/Models/UserModels/CorporateUser.cs
@@ -4,7 +4,7 @@ namespace AuctionData.Models.UserModels;
 
 public class CorporateUser : User
 {
-    public long Credit { get; set; }
+    public decimal Credit { get; set; }
     public string? CvrNumber { get; set; }
     public int UserId { get; set; }
     public User? User { get; set; }
